+++ src/elf/mod.rs
@@ -300,7 +300,7 @@ if_sylvan! {
 
             let mut syms = Symtab::default();
             let mut strtab = Strtab::default();
-            if let Some(shdr) = section_headers.iter().rfind(|shdr| shdr.sh_type as u32 == section_header::SHT_SYMTAB) {
+            if let Some(shdr) = section_headers.iter().rfind(|shdr| shdr.sh_type == section_header::SHT_SYMTAB) {
                 let size = shdr.sh_entsize;
                 let count = if size == 0 { 0 } else { shdr.sh_size / size };
                 syms = Symtab::parse(bytes, shdr.sh_offset as usize, count as usize, ctx)?;
@@ -345,7 +345,7 @@ if_sylvan! {
                 // parse the dynamic relocations
                 dynrelas = RelocSection::parse(bytes, dyn_info.rela, dyn_info.relasz, true, ctx)?;
                 dynrels = RelocSection::parse(bytes, dyn_info.rel, dyn_info.relsz, false, ctx)?;
-                let is_rela = dyn_info.pltrel as u64 == dynamic::DT_RELA;
+                let is_rela = dyn_info.pltrel == dynamic::DT_RELA;
                 pltrelocs = RelocSection::parse(bytes, dyn_info.jmprel, dyn_info.pltrelsz, is_rela, ctx)?;
 
                 let mut num_syms = if let Some(gnu_hash) = dyn_info.gnu_hash {
